+++ selenium_test/steps/saving.py
@@ -50,7 +50,7 @@ This paragraph and its content are designed to test how error markers \
 are shown for inline elements that end up spanning multiple lines. \
 This paragraph and its content are designed to test how error markers \
 are shown for inline elements that end up spanning multiple lines.\
-</monogr></p></body></text></TEI>""",
+</monogr></p><p n="">P</p></body></text></TEI>""",
     "serializes multiple top namespaces properly": """\
 <TEI xmlns="http://www.tei-c.org/ns/1.0" \
 xmlns:math="http://www.w3.org/1998/Math/MathML"><teiHeader><fileDesc>\